// src/app/lpb-services/lpb-credit-card-service/shared/models/credit-card.ts
export interface UserDetails {
  completed?: boolean;
  id: string;
  approvalDateLos: string;
  approveLv1By?: string;
  approveLv2By?: string;
  assetCode: string;
  birthDate: string;
  branchCode: string;
  cardCategory: string;
  cardHolderName: string;
  cardLosCode: string;
  cardRateCode: string;
  cardTypeName: string;
  city: string;
  createdDate?: string;
  creditCardKey: string;
  custCatType: string;
  customerCategory: string;
  customerCode: string;
  debtAccount: string;
  debtAmount: string;
  displayStatus: string;
  email: string;
  employeeId: string;
  gender: string;
  getLosDate?: string;
  houseNumber: string;
  inviteCode: string;
  issueDate: string;
  issuePlace: string;
  issueType: string;
  issueValue: string;
  lastName: string;
  limitAmount: number;
  note?: string;
  phoneNumber: string;
  processStatus: string;
  productCode: string;
  productId: string;
  productType: string;
  rejectReason?: string;
  secretAnswer: string;
  secretQuestion: string;
  statusLos: string;
  street: string;
  surname: string;
}